+++ cpp/src/io/shp/legacy/polygon_shapefile_reader.cpp
@@ -25,7 +25,7 @@
 #include <cudf/legacy/column.hpp>
 #include <cudf/utilities/error.hpp>
 #include <cuspatial/legacy/shapefile_readers.hpp>
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 
 #include <ogrsf_frmts.h>