+++ cpp/src/trajectory/legacy/trajectory_spatial_bound.cu
@@ -18,7 +18,7 @@
 #include <utilities/legacy/cuda_utils.hpp>
 #include <type_traits>
 
-#include <utility/utility.hpp>
+#include <utility/legacy/utility.hpp>
 #include <utility/trajectory_thrust.cuh>
 #include <cuspatial/legacy/trajectory.hpp>
 #include <cuspatial/error.hpp>