+++ cpp/tests/io/legacy/read_shapefile_polygon_test.cu
@@ -23,7 +23,7 @@
 #include <tests/utilities/legacy/cudf_test_utils.cuh>
 #include <tests/utilities/legacy/cudf_test_fixtures.h>
 #include <cuspatial/legacy/shapefile_readers.hpp>
-#include <utility/utility.hpp> 
+#include <utility/legacy/utility.hpp>
 
 
 struct ReadShapefilePolygonTest : public GdfTest