+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - PR #163 Fix cudf legacy Cython imports/cimports
 - PR #166 Move trajectory.hpp files to legacy
 - PR #167 Align utility.hpp with libcudf style
+- PR #171 Update trajectory.hpp to libcudf++
 - PR #173 Move hausdorff.hpp files to legacy
 - PR #172 Move coordinate_transform.hpp files to legacy
 - PR #170 Update coordinate_transform.hpp to libcudf++
@@ -18,7 +19,7 @@
 - PR #190 Port coordinate_transform.hpp cython files
 - PR #191 Move point_in_polygon.hpp files to legacy
 - PR #193 Move shapefile_readers.hpp files to legacy
-- PR #171 Update trajectory.hpp to libcudf++
+- PR #196 Move utilities/utilities.hpp to legacy
 
 ## Bug Fixes